+++ source/DSCResources/DSC_SqlMemory/DSC_SqlMemory.psm1
@@ -85,8 +85,16 @@ function Get-TargetResource
         This is the minimum amount of memory, in MB, in the buffer pool used by
         the instance of SQL Server.
 
+    .PARAMETER MinMemoryPercent
+        This is the minimum amount of memory, as a percentage of total server memory, in the buffer pool used by
+        the instance of SQL Server.
+
     .PARAMETER MaxMemory
         This is the maximum amount of memory, in MB, in the buffer pool used by
+        the instance of SQL Server.
+
+    .PARAMETER MaxMemoryPercent
+        This is the maximum amount of memory, as a percentage of total server memory, in the buffer pool used by
         the instance of SQL Server.
 
     .PARAMETER ProcessOnlyOnActiveNode
@@ -124,8 +132,18 @@ function Set-TargetResource
         $MinMemory,
 
         [Parameter()]
+        [ValidateRange(1, 100)]
+        [System.Int32]
+        $MinMemoryPercent,
+
+        [Parameter()]
         [System.Int32]
         $MaxMemory,
+
+        [Parameter()]
+        [ValidateRange(1, 100)]
+        [System.Int32]
+        $MaxMemoryPercent,
 
         [Parameter()]
         [System.Boolean]
@@ -151,6 +169,12 @@ function Set-TargetResource
                         New-InvalidArgumentException -ArgumentName 'MaxMemory' -Message $errorMessage
                     }
 
+                    if ($MaxMemoryPercent)
+                    {
+                        $errorMessage = $script:localizedData.MaxMemoryPercentParamMustBeNull
+                        New-InvalidArgumentException -ArgumentName 'MaxMemoryPercent' -Message $errorMessage
+                    }
+
                     $MaxMemory = Get-SqlDscDynamicMaxMemory
 
                     Write-Verbose -Message (
@@ -159,21 +183,56 @@ function Set-TargetResource
                 }
                 else
                 {
-                    if (-not $MaxMemory)
+                    if (-not $MaxMemory -and -not $MaxMemoryPercent)
                     {
                         $errorMessage = $script:localizedData.MaxMemoryParamMustNotBeNull
                         New-InvalidArgumentException -ArgumentName 'MaxMemory' -Message $errorMessage
                     }
                 }
 
-                Write-Verbose -Message (
-                    $script:localizedData.MaximumMemoryLimited -f $InstanceName, $MaxMemory
-                )
+                if ($MaxMemory)
+                {
+                    if ($MaxMemoryPercent)
+                    {
+                        $errorMessage = $script:localizedData.MaxMemoryPercentParamMustBeNull
+                        New-InvalidArgumentException -ArgumentName 'MaxMemoryPercent' -Message $errorMessage
+                    }
 
-                $sqlServerObject.Configuration.MaxServerMemory.ConfigValue = $MaxMemory
+                    $sqlServerObject.Configuration.MaxServerMemory.ConfigValue = $MaxMemory
+
+                    Write-Verbose -Message (
+                        $script:localizedData.MaximumMemoryLimited -f $InstanceName, $MaxMemory
+                    )
+                }
+                elseif ($MaxMemoryPercent)
+                {
+                    $MaxMemory = Get-SqlDscPercentMemory -PercentMemory $MaxMemoryPercent
+
+                    $sqlServerObject.Configuration.MaxServerMemory.ConfigValue = $MaxMemory
+
+                    Write-Verbose -Message (
+                        $script:localizedData.MaximumMemoryLimited -f $InstanceName, $MaxMemory
+                    )
+                }
 
                 if ($MinMemory)
                 {
+                    if ($MinMemoryPercent)
+                    {
+                        $errorMessage = $script:localizedData.MinMemoryPercentParamMustBeNull
+                        New-InvalidArgumentException -ArgumentName 'MinMemoryPercent' -Message $errorMessage
+                    }
+
+                    $sqlServerObject.Configuration.MinServerMemory.ConfigValue = $MinMemory
+
+                    Write-Verbose -Message (
+                        $script:localizedData.MinimumMemoryLimited -f $InstanceName, $MinMemory
+                    )
+                }
+                elseif ($MinMemoryPercent)
+                {
+                    $MinMemory = Get-SqlDscPercentMemory -PercentMemory $MinMemoryPercent
+
                     $sqlServerObject.Configuration.MinServerMemory.ConfigValue = $MinMemory
 
                     Write-Verbose -Message (
@@ -238,8 +297,16 @@ function Set-TargetResource
         This is the minimum amount of memory, in MB, in the buffer pool used by
         the instance of SQL Server.
 
+    .PARAMETER MinMemoryPercent
+        This is the minimum amount of memory, as a percentage of total server memory, in the buffer pool used by
+        the instance of SQL Server.
+
     .PARAMETER MaxMemory
         This is the maximum amount of memory, in MB, in the buffer pool used by
+        the instance of SQL Server.
+
+    .PARAMETER MaxMemoryPercent
+        This is the maximum amount of memory, as a percentage of total server memory, in the buffer pool used by
         the instance of SQL Server.
 
     .PARAMETER ProcessOnlyOnActiveNode
@@ -277,8 +344,18 @@ function Test-TargetResource
         $MinMemory,
 
         [Parameter()]
+        [ValidateRange(1, 100)]
+        [System.Int32]
+        $MinMemoryPercent,
+
+        [Parameter()]
         [System.Int32]
         $MaxMemory,
+
+        [Parameter()]
+        [ValidateRange(1, 100)]
+        [System.Int32]
+        $MaxMemoryPercent,
 
         [Parameter()]
         [System.Boolean]
@@ -346,6 +423,12 @@ function Test-TargetResource
                     New-InvalidArgumentException -ArgumentName 'MaxMemory' -Message $errorMessage
                 }
 
+                if ($MaxMemoryPercent)
+                {
+                    $errorMessage = $script:localizedData.MaxMemoryPercentParamMustBeNull
+                    New-InvalidArgumentException -ArgumentName 'MaxMemoryPercent' -Message $errorMessage
+                }
+
                 $MaxMemory = Get-SqlDscDynamicMaxMemory
 
                 Write-Verbose -Message (
@@ -354,11 +437,24 @@ function Test-TargetResource
             }
             else
             {
-                if (-not $MaxMemory)
+                if (-not $MaxMemory -and -not $MaxMemoryPercent)
                 {
                     $errorMessage = $script:localizedData.MaxMemoryParamMustNotBeNull
                     New-InvalidArgumentException -ArgumentName 'MaxMemory' -Message $errorMessage
                 }
+            }
+
+            if ($MaxMemory)
+            {
+                if ($MaxMemoryPercent)
+                {
+                    $errorMessage = $script:localizedData.MaxMemoryPercentParamMustBeNull
+                    New-InvalidArgumentException -ArgumentName 'MaxMemoryPercent' -Message $errorMessage
+                }
+            }
+            elseif ($MaxMemoryPercent)
+            {
+                $MaxMemory = Get-SqlDscPercentMemory -PercentMemory $MaxMemoryPercent
             }
 
             if ($MaxMemory -ne $currentMaxMemory)
@@ -370,8 +466,21 @@ function Test-TargetResource
                 $isServerMemoryInDesiredState = $false
             }
 
-            if ($MinMemory)
+            if ($MinMemory -or $MinMemoryPercent)
             {
+                if ($MinMemory)
+                {
+                    if ($MinMemoryPercent)
+                    {
+                        $errorMessage = $script:localizedData.MinMemoryPercentParamMustBeNull
+                        New-InvalidArgumentException -ArgumentName 'MinMemoryPercent' -Message $errorMessage
+                    }
+                }
+                elseif ($MinMemoryPercent)
+                {
+                    $MinMemory = Get-SqlDscPercentMemory -PercentMemory $MinMemoryPercent
+                }
+
                 if ($MinMemory -ne $currentMinMemory)
                 {
                     Write-Verbose -Message (
@@ -447,3 +556,34 @@ function Get-SqlDscDynamicMaxMemory
     $maxMemory
 }
 
+<#
+    .SYNOPSIS
+        This function returns the amount of memory in MB, calculated from the input percentage of total server memory.
+
+    .PARAMETER MemoryPercent
+        This is the percentage of total server memory to calculate.
+#>
+function Get-SqlDscPercentMemory
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateRange(1, 100)]
+        [System.Int32]
+        $PercentMemory
+    )
+
+    try
+    {
+        $physicalMemory = (Get-CimInstance -ClassName Win32_ComputerSystem).TotalPhysicalMemory
+        $memoryInMegaBytes = [Math]::Round(($physicalMemory * ($PercentMemory/100)) / 1MB)
+    }
+    catch
+    {
+        $errorMessage = $script:localizedData.ErrorGetPercentMemory
+        New-InvalidOperationException -Message $errorMessage -ErrorRecord $_
+    }
+
+    $memoryInMegaBytes
+}